Master: Add Task SD and configure SD Card
Adding task base SD Card, working in progress.
Configure SD module IO Pin and clock divider (STM32Duino modified default L4 time).
Unified module CAN/SENSOR config.
Adding UID CPU and connect to serialnumber and CAN UniquieID PnP request/response.
Setup PIN Mode Arduino in module_hal (working in progress)
Report in slave UID and SerialNumber.
Removed ununsed boot_loader struct field.

+++ platformio/stima_v4/slave-th/src/canard_class_th.cpp
@@ -769,12 +769,12 @@ bool canardClass::slave_heartbeat_send_message(void)
 /// @brief Invia il messaggio di PNP request (richiesta di node_id valido) al nodo server PNP (master)
 /// @param  None
 /// @return true se il metodo è eseguito correttamente
-bool canardClass::slave_pnp_send_request(uint32_t serial_number) {
+bool canardClass::slave_pnp_send_request(uint64_t serial_number) {
     // PnP over Classic CAN, use message v1.0.
     uavcan_pnp_NodeIDAllocationData_1_0 msg = {0};
     // truncated uint48 unique_id_hash
     // msg.allocated_node_id.(count/element) => Solo in response non in request;
-    msg.unique_id_hash = (uint64_t)serial_number;
+    msg.unique_id_hash = serial_number;
     uint8_t serialized[uavcan_pnp_NodeIDAllocationData_1_0_SERIALIZATION_BUFFER_SIZE_BYTES_] = {0};
     size_t serialized_size = sizeof(serialized);
     const int8_t err = uavcan_pnp_NodeIDAllocationData_1_0_serialize_(&msg, &serialized[0], &serialized_size);
